hw2: rename package
