+++ hw1-student-testing/src/main/resources/spring-context.xml
@@ -4,7 +4,7 @@
        xsi:schemaLocation="http://www.springframework.org/schema/beans
         http://www.springframework.org/schema/beans/spring-beans.xsd">
 
-    <bean id="parser" class="ru.otus.spring.parser.OpenCsvParser">
+    <bean id="parser" class="ru.otus.spring.hw1.parser.OpenCsvParser">
         <property name="fileName" value="test.csv"/>
     </bean>
