--- hw2-student-testing/src/main/resources/.~lock.test.csv#
@@ -1,1 +1,0 @@
-,alsu,alsu-Yoga-Slim-7-Pro-14ACH5,01.10.2022 16:46,file:///home/alsu/.config/libreoffice/4;
